Auto change - after each gyro turn, a check and adjustment is done to ensure turn was made within a degree of error beyond that of the heading threshold to account for overshooting target. Disabled unnecessary programs. Ready for Worlds.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Auto_Master.java
@@ -63,7 +63,7 @@ public abstract class Auto_Master extends LinearOpMode {
 
     RelicRecoveryVuMark vuMark;
 
-    int angleAdjust = 0;
+    double angleAdjust = 0;
     int distAdjust = 0;
 
     int left1Pos;
